requirements.txt
requests
beautifulsoup4
lxml
sentence-transformers
numpy
qdrant-client
tqdm
click
fastapi
uvicorn[standard]
python-dotenv
google-genai
authlib
httpx
PyJWT
itsdangerous
